+++ sdk/search/arm-search/src/searchManagementClient.ts
@@ -21,8 +21,7 @@ import {
   ServicesImpl,
   PrivateLinkResourcesImpl,
   PrivateEndpointConnectionsImpl,
-  SharedPrivateLinkResourcesImpl,
-  UsagesImpl
+  SharedPrivateLinkResourcesImpl
 } from "./operations";
 import {
   Operations,
@@ -31,16 +30,9 @@ import {
   Services,
   PrivateLinkResources,
   PrivateEndpointConnections,
-  SharedPrivateLinkResources,
-  Usages
+  SharedPrivateLinkResources
 } from "./operationsInterfaces";
-import * as Parameters from "./models/parameters";
-import * as Mappers from "./models/mappers";
-import {
-  SearchManagementClientOptionalParams,
-  UsageBySubscriptionSkuOptionalParams,
-  UsageBySubscriptionSkuResponse
-} from "./models";
+import { SearchManagementClientOptionalParams } from "./models";
 
 export class SearchManagementClient extends coreClient.ServiceClient {
   $host: string;
@@ -75,7 +67,7 @@ export class SearchManagementClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-search/3.2.0`;
+    const packageDetails = `azsdk-js-arm-search/4.0.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -128,7 +120,7 @@ export class SearchManagementClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-11-01";
+    this.apiVersion = options.apiVersion || "2021-04-01-preview";
     this.operations = new OperationsImpl(this);
     this.adminKeys = new AdminKeysImpl(this);
     this.queryKeys = new QueryKeysImpl(this);
@@ -136,7 +128,6 @@ export class SearchManagementClient extends coreClient.ServiceClient {
     this.privateLinkResources = new PrivateLinkResourcesImpl(this);
     this.privateEndpointConnections = new PrivateEndpointConnectionsImpl(this);
     this.sharedPrivateLinkResources = new SharedPrivateLinkResourcesImpl(this);
-    this.usages = new UsagesImpl(this);
     this.addCustomApiVersionPolicy(options.apiVersion);
   }
 
@@ -168,23 +159,6 @@ export class SearchManagementClient extends coreClient.ServiceClient {
     this.pipeline.addPolicy(apiVersionPolicy);
   }
 
-  /**
-   * Gets the quota usage for a search sku in the given subscription.
-   * @param location The unique location name for a Microsoft Azure geographic region.
-   * @param skuName The unique search service sku name supported by Azure Cognitive Search.
-   * @param options The options parameters.
-   */
-  usageBySubscriptionSku(
-    location: string,
-    skuName: string,
-    options?: UsageBySubscriptionSkuOptionalParams
-  ): Promise<UsageBySubscriptionSkuResponse> {
-    return this.sendOperationRequest(
-      { location, skuName, options },
-      usageBySubscriptionSkuOperationSpec
-    );
-  }
-
   operations: Operations;
   adminKeys: AdminKeys;
   queryKeys: QueryKeys;
@@ -192,30 +166,4 @@ export class SearchManagementClient extends coreClient.ServiceClient {
   privateLinkResources: PrivateLinkResources;
   privateEndpointConnections: PrivateEndpointConnections;
   sharedPrivateLinkResources: SharedPrivateLinkResources;
-  usages: Usages;
 }
-// Operation Specifications
-const serializer = coreClient.createSerializer(Mappers, /* isXml */ false);
-
-const usageBySubscriptionSkuOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/subscriptions/{subscriptionId}/providers/Microsoft.Search/locations/{location}/usages/{skuName}",
-  httpMethod: "GET",
-  responses: {
-    200: {
-      bodyMapper: Mappers.QuotaUsageResult
-    },
-    default: {
-      bodyMapper: Mappers.CloudError
-    }
-  },
-  queryParameters: [Parameters.apiVersion],
-  urlParameters: [
-    Parameters.$host,
-    Parameters.subscriptionId,
-    Parameters.location,
-    Parameters.skuName
-  ],
-  headerParameters: [Parameters.accept, Parameters.clientRequestId],
-  serializer
-};